FIX : Réparation des Mappers : suppression de la librairie MapStruct, cassant le fonctionnement des TI JpaIntegrationTests, utilisation de mappers classiques

+++ src/main/java/papillon/colibri/corbeau/entities/SauterelleEntity.java
@@ -17,12 +17,11 @@ public class SauterelleEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "ID", nullable = false)
-    private Byte id;
+    private Integer id;
 
     @Column(name = "COULEUR", nullable = false, length = 50)
     private String couleur;
 
     @Column(name = "NAISSANCE")
     private LocalDate naissance;
-
 }
